add tests for the validation methods
