+++ src/commands/sbom/upload.ts
@@ -3,7 +3,6 @@ import os from 'os'
 import process from 'process'
 
 import Ajv from 'ajv'
-import addFormats from 'ajv-formats'
 import {AxiosPromise, AxiosResponse} from 'axios'
 import chalk from 'chalk'
 import {Command, Option} from 'clipanion'
@@ -12,55 +11,9 @@ import {SpanTags} from '../../helpers/interfaces'
 import {getSpanTags} from '../../helpers/tags'
 
 import {getApiHelper} from './api'
-import cycloneDxSchema from './json-schema/cyclonedx/bom-1.4.schema.json'
-import jsfSchema from './json-schema/jsf/jsf-0.82.schema.json'
-import spdxSchema from './json-schema/spdx/spdx.schema.json'
 import {SBOMEntity, SBOMPayload, SBOMSourceType} from './protobuf/sbom_intake'
 import {SbomPayloadData} from './types'
-
-/**
- * Get the validate function. Read all the schemas and return
- * the function used to validate all SBOM documents.
- */
-const getValidator = (): Ajv => {
-  const ajv = new Ajv({strict: false, validateFormats: false})
-  ajv.addMetaSchema(spdxSchema)
-  ajv.addMetaSchema(jsfSchema)
-  addFormats(ajv)
-
-  return ajv
-}
-
-/**
- * Validate an SBOM file.
- * @param path - the path of the file to validate
- * @param ajv - an instance of Ajv fully initialized and ready to use.
- */
-const validateSbomFile = (path: string, ajv: Ajv): boolean => {
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
-    const fileContent = JSON.parse(String(fs.readFileSync(path)))
-    const validateFunction = ajv.compile(cycloneDxSchema)
-
-    const isValid = validateFunction(fileContent)
-
-    if (!isValid) {
-      const errors = validateFunction.errors || []
-
-      errors.forEach((em) => {
-        process.stderr.write(`Error while reading file: ${em}\n`)
-      })
-
-      return false
-    }
-
-    return true
-  } catch (error) {
-    process.stderr.write(`Error while reading file: ${error.message}\n`)
-
-    return false
-  }
-}
+import {getValidator, validateSbomFile} from './validation'
 
 const generatePayload = (payloadData: SbomPayloadData, tags: SpanTags): SBOMPayload => {
   const spanTagsAsStringArray = Object.keys(tags).map((key) => `${key}:${tags[key as keyof SpanTags]}`)
